#805 fix(bash): preserve adapter layering in copied routes
